Split sm_daemons and move daemon-related code into a submodule
