+++ metaspace/engine/scripts/run_sm_daemon.py
@@ -8,7 +8,9 @@ from sm.engine.daemons.lithops_daemon import LithopsDaemon
 from sm.engine.db import DB, ConnectionPool
 from sm.engine.es_export import ESExporter
 from sm.engine.image_store import ImageStoreServiceWrapper
-from sm.engine.sm_daemons import SMAnnotateDaemon, DatasetManager, SMIndexUpdateDaemon
+from sm.engine.daemons.update_daemon import SMIndexUpdateDaemon
+from sm.engine.daemons.annotate_daemon import SMAnnotateDaemon
+from sm.engine.daemons.dataset_manager import DatasetManager
 from sm.engine.queue import (
     SM_ANNOTATE,
     SM_UPDATE,
