+++ metaspace/engine/sm/engine/daemons/dataset_manager.py
@@ -1,33 +1,27 @@
 import json
 import logging
-from traceback import format_exc
 import urllib.parse
-from pathlib import Path
 
 import boto3
 from botocore.exceptions import ClientError
-import redis
-from requests import post
+from requests.api import post
 
+from sm.engine import molecular_db
 from sm.engine.annotation.job import del_jobs
 from sm.engine.annotation_lithops.annotation_job import ServerAnnotationJob
 from sm.engine.annotation_lithops.executor import Executor
+from sm.engine.annotation_spark.annotation_job import AnnotationJob
 from sm.engine.colocalization import Colocalization
-from sm.engine.daemon_action import DaemonAction, DaemonActionStage
+from sm.engine.daemons.daemon_action import DaemonActionStage
 from sm.engine.dataset import Dataset, DatasetStatus
 from sm.engine.db import DB
-from sm.engine.errors import AnnotationError, ImzMLError, IndexUpdateError, SMError, UnknownDSID
 from sm.engine.es_export import ESExporter
 from sm.engine.ion_thumbnail import generate_ion_thumbnail, generate_ion_thumbnail_lithops
 from sm.engine.isocalc_wrapper import IsocalcWrapper
-from sm.engine import molecular_db
 from sm.engine.off_sample_wrapper import classify_dataset_ion_images
 from sm.engine.optical_image import del_optical_image
-from sm.engine.queue import QueueConsumer, QueuePublisher
 from sm.engine.util import SMConfig
 from sm.engine.utils.perf_profile import perf_profile
-from sm.rest.dataset_manager import DatasetActionPriority
-from sm.engine.annotation_spark.annotation_job import AnnotationJob
 
 
 class DatasetManager:
@@ -223,181 +217,3 @@ class DatasetManager:
         self.notify_update(ds.id, msg['action'], stage=DaemonActionStage.FAILED)
         slack_msg = f'{json.dumps(msg)}\n```{e.traceback}```'
         self.post_to_slack('hankey', f" [x] {msg['action']} failed: {slack_msg}")
-
-
-class SMAnnotateDaemon:
-    """ Reads messages from annotation queue and starts annotation jobs
-    """
-
-    logger = logging.getLogger('annotate-daemon')
-
-    def __init__(self, manager, annot_qdesc, upd_qdesc, poll_interval=1):
-        self._sm_config = SMConfig.get_conf()
-        self._stopped = False
-        self._manager = manager
-        self._annot_queue_consumer = QueueConsumer(
-            config=self._sm_config['rabbitmq'],
-            qdesc=annot_qdesc,
-            callback=self._callback,
-            on_success=self._on_success,
-            on_failure=self._on_failure,
-            logger=self.logger,
-            poll_interval=poll_interval,
-        )
-        self._update_queue_pub = QueuePublisher(
-            config=self._sm_config['rabbitmq'], qdesc=upd_qdesc, logger=self.logger
-        )
-        self._redis_client = redis.Redis(**self._sm_config.get('redis', {}))
-        Path(self._sm_config['fs']['spark_data_path']).mkdir(parents=True, exist_ok=True)
-
-    def _on_success(self, msg):
-        self.logger.info(f" SM annotate daemon: success")
-
-        ds = self._manager.load_ds(msg['ds_id'])
-        self._manager.set_ds_status(ds, DatasetStatus.FINISHED)
-        self._manager.notify_update(ds.id, msg['action'], DaemonActionStage.FINISHED)
-
-        self._manager.post_to_slack('dart', ' [v] Annotation succeeded: {}'.format(json.dumps(msg)))
-        self._redis_client.set('cluster-busy', 'no')
-
-    def _on_failure(self, msg, e):
-        self._manager.ds_failure_handler(msg, e)
-
-        if 'email' in msg:
-            traceback = e.__cause__.traceback if isinstance(e.__cause__, ImzMLError) else None
-            self._manager.send_failed_email(msg, traceback)
-        self._redis_client.set('cluster-busy', 'no')
-
-    def _callback(self, msg):
-        try:
-            self.logger.info(f" SM annotate daemon received a message: {msg}")
-            self._redis_client.set('cluster-busy', 'yes', ex=3600 * 13)  # key expires in 13h
-
-            ds = self._manager.load_ds(msg['ds_id'])
-            self._manager.set_ds_status(ds, DatasetStatus.ANNOTATING)
-            self._manager.notify_update(ds.id, msg['action'], DaemonActionStage.STARTED)
-
-            self._manager.post_to_slack(
-                'new', " [v] New annotation message: {}".format(json.dumps(msg))
-            )
-
-            self._manager.annotate(ds=ds, del_first=msg.get('del_first', False))
-
-            update_msg = {
-                'ds_id': msg['ds_id'],
-                'ds_name': msg['ds_name'],
-                'email': msg.get('email', None),
-                'action': DaemonAction.INDEX,
-            }
-            self._update_queue_pub.publish(msg=update_msg, priority=DatasetActionPriority.HIGH)
-
-            if self._sm_config['services'].get('off_sample', False):
-                analyze_msg = {
-                    'ds_id': msg['ds_id'],
-                    'ds_name': msg['ds_name'],
-                    'action': DaemonAction.CLASSIFY_OFF_SAMPLE,
-                }
-                self._update_queue_pub.publish(msg=analyze_msg, priority=DatasetActionPriority.LOW)
-        except Exception as e:
-            raise AnnotationError(ds_id=msg['ds_id'], traceback=format_exc(chain=False)) from e
-
-    def start(self):
-        self._stopped = False
-        self._annot_queue_consumer.start()
-
-    def stop(self):
-        """  Must be called from main thread
-        """
-        if not self._stopped:
-            self._annot_queue_consumer.stop()
-            self._annot_queue_consumer.join()
-            self._stopped = True
-
-
-class SMIndexUpdateDaemon:
-    """ Reads messages from the update queue and does indexing/update/delete
-    """
-
-    logger = logging.getLogger('update-daemon')
-
-    def __init__(self, manager, make_update_queue_cons):
-        self._manager = manager
-        self._update_queue_cons = make_update_queue_cons(
-            callback=self._callback, on_success=self._on_success, on_failure=self._on_failure
-        )
-        self._stopped = False
-
-    def _on_success(self, msg):
-        self.logger.info(f" SM update daemon: success")
-
-        if msg['action'] == DaemonAction.DELETE:
-            self._manager.notify_update(
-                msg['ds_id'], action=DaemonAction.DELETE, stage=DaemonActionStage.FINISHED
-            )
-        else:
-            ds = self._manager.load_ds(msg['ds_id'])
-            if msg['action'] == DaemonAction.INDEX:
-                self._manager.set_ds_status(ds, DatasetStatus.FINISHED)
-            self._manager.notify_update(ds.id, msg['action'], DaemonActionStage.FINISHED)
-
-        if msg['action'] in [DaemonAction.UPDATE, DaemonAction.INDEX]:
-            msg['web_app_link'] = self._manager.create_web_app_link(msg)
-
-        if msg['action'] != DaemonAction.UPDATE:
-            self._manager.post_to_slack(
-                'dart', f" [v] Succeeded to {msg['action']}: {json.dumps(msg)}"
-            )
-
-        if msg.get('email'):
-            self._manager.send_success_email(msg)
-
-    def _on_failure(self, msg, e):
-        self._manager.ds_failure_handler(msg, e)
-
-        if 'email' in msg:
-            self._manager.send_failed_email(msg)
-
-    def _callback(self, msg):
-        try:
-            self.logger.info(f' SM update daemon received a message: {msg}')
-            self._manager.post_to_slack(
-                'new', f" [v] New {msg['action']} message: {json.dumps(msg)}"
-            )
-
-            ds = self._manager.load_ds(msg['ds_id'])
-            self._manager.notify_update(ds.id, msg['action'], DaemonActionStage.STARTED)
-
-            if msg['action'] == DaemonAction.INDEX:
-                self._manager.index(ds=ds)
-
-            elif msg['action'] == DaemonAction.CLASSIFY_OFF_SAMPLE:
-                try:
-                    # depending on number of annotations may take up to several minutes
-                    self._manager.classify_dataset_images(ds)
-                except Exception as e:  # don't fail dataset when off-sample pred fails
-                    self.logger.warning(f'Failed to classify off-sample: {e}')
-
-                try:
-                    self._manager.index(ds=ds)
-                except UnknownDSID:
-                    # Sometimes the DS will have been deleted before this point
-                    self.logger.warning(f'DS missing after off-sample classification: {ds.id}')
-
-            elif msg['action'] == DaemonAction.UPDATE:
-                self._manager.update(ds, msg['fields'])
-            elif msg['action'] == DaemonAction.DELETE:
-                self._manager.delete(ds=ds)
-            else:
-                raise SMError(f"Wrong action: {msg['action']}")
-        except Exception as e:
-            raise IndexUpdateError(msg['ds_id'], traceback=format_exc(chain=False)) from e
-
-    def start(self):
-        self._stopped = False
-        self._update_queue_cons.start()
-
-    def stop(self):
-        if not self._stopped:
-            self._update_queue_cons.stop()
-            self._update_queue_cons.join()
-            self._stopped = True
