+++ metaspace/engine/sm/rest/dataset_manager.py
@@ -10,7 +10,7 @@ from sm.engine.dataset import (
     FLAT_DS_CONFIG_KEYS,
 )
 from sm.engine.errors import DSIsBusy, UnknownDSID
-from sm.engine.daemon_action import DaemonAction, DaemonActionStage
+from sm.engine.daemons.daemon_action import DaemonAction, DaemonActionStage
 from sm.engine.optical_image import add_optical_image, del_optical_image
 from sm.engine.util import SMConfig
