+++ metaspace/engine/tests/test_sm_daemons.py
@@ -15,7 +15,7 @@ from fabric.context_managers import warn_only
 import numpy as np
 import pandas as pd
 
-from sm.engine.daemon_action import DaemonAction
+from sm.engine.daemons.daemon_action import DaemonAction
 from sm.engine.db import DB
 from sm.engine.es_export import ESExporter
 from sm.engine.dataset import DatasetStatus
@@ -90,7 +90,9 @@ def queue_pub(local_sm_config):
 def run_daemons(db, es, sm_config):
     from sm.engine.queue import QueuePublisher, SM_DS_STATUS, SM_ANNOTATE, SM_UPDATE
     from sm.engine.image_store import ImageStoreServiceWrapper
-    from sm.engine.sm_daemons import DatasetManager, SMAnnotateDaemon, SMIndexUpdateDaemon
+    from sm.engine.daemons.dataset_manager import DatasetManager
+    from sm.engine.daemons.annotate_daemon import SMAnnotateDaemon
+    from sm.engine.daemons.update_daemon import SMIndexUpdateDaemon
 
     status_queue_pub = QueuePublisher(
         config=sm_config['rabbitmq'], qdesc=SM_DS_STATUS, logger=logger
